Workaround the two mac startup issues
1) Error message: unable to open documentviewcontroller because view
is not in view hierarchy.  That is because the diagramviewcontroller
is already open and we are trying to present it again.
2) DiagramViewController is no longer root view controller when
opening modal dialogs, like the About dialog.  We set the diagram view
controller as root view controller to avoid this, and apparently when
we call the document browser view controller again, it goes back to
being the root view controller.